2021 Week 3 | Power BI: Report Interactivity
Requirements
Add a slicer for [Year] to the report. 
Edit Interactions for the slicer to have it only slice the KPIs and bar charts (not trends)
Edit Interactions of the Conference breakdowns so that they filter the KPIs and Trends on click. 
Use colors to differentiate between metrics
Figure out how much money the MAC conference made in 2016
Dataset
This week uses a data set that breaks down NCAA athletic department expenses and revenues by year. You can get it here at data.world.