2021 Week 10 | Power BI: Violin Plot Custom Visual
Requirements
We are using a new dataset this week. Please download the starter Power BI file here.
Create a violin plot that displays average temperature grouped by month name. The average temperature value is provided in both Fahrenheit (the field named Avg Temp F) and Celsius (the field named Avg Temp C). You can use either field according to your preference. The example and solution video use the Avg Temp F field. 
Use the barcode (strip) plot to show the individual temperature values. 
Use the Epanechnikov (default) kernel and a high sampling resolution for the violin shape. 
Show lines in the plot for the first and third quartiles and the median.
Use a text size of at least 12 points for the x-axis and y-axis labels.
Format the tooltip so that the measures display as integers (show zero decimal places). 
Title the chart “2020 Daily Average Temperatures in Denver, Colorado”
You may choose your own colors, but you must ensure that you have a color contrast ratio of 3:1 between the bars for the individual data points and the kernel density background, between the reference lines for the quartiles and median and the kernel density background, and between the individual data points and the reference lines for the quartiles and median. The WebAIM Contrast Checker or Colour Contrast Analyser may be helpful for this requirement. 
Identify the months with the highest and lowest standard deviation.